requirements.txt
agora-api==0.25.1
aiohttp==3.5.4
aiohttp-sse-client==0.1.4
async-timeout==3.0.1
attrs==21.2.0
base58==2.0.1
cffi==1.14.5
chardet==3.0.4
crc16==0.1.1
ed25519==1.4
grpcio==1.34.1
idna==3.1
kin-base==1.4.1
kin-sdk-v2==0.6.1
mnemonic==0.18
multidict==4.7.6
pbkdf2==1.3
protobuf==3.12.2
pycparser==2.20
PyNaCl==1.4.0
six==1.16.0
typing-extensions==3.10.0.0
yarl==1.6.3
